# tls: Prefer faster ciphers and use server preference