@StephanDollberg 2023-06-05
> the other question i had was: in the gnutls docs it says that this setting causes the server's priorities to take precedence. does this mean that negotiation with a client could fail if the server's high priority setting was incompatible with the clients? this means sense for our internal traffic, but a little more concerning if it is a kafka client. anything to be concerned about here regarding that?

No that shouldn't be an issue as long as we still offer the same ciphers (which we do as per above).

The server always choses the cipher but with server preference it goes by its own list in priority order instead of the clients. 

For example assume the server has the supported list (in priority order) A B C and the client has D C A.

B/D would never be picked as not supported by the client/server. With client preference the server would chose C as it's supported by both the client and the server and is first in the client priority list. With server preference the server would pick A as it's first in the server priority list and supported by both the client and the server. 

Does that make sense?